Move property/3 to new API
